requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg2==1.19.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14